requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
jinja2>=3.1.4
networkx>=3.3
pydantic>=2.7.0
python-multipart>=0.0.9
httpx>=0.27.0
google-generativeai>=0.8.0
